requirements.txt
openai~=1.24.0
python-dotenv~=1.0.1
pdfplumber~=0.11.0
langchain~=0.1.17
anthropic~=0.25.7
requests~=2.31.0
langchain-openai
unstructured
unstructured[pdf]
streamlit~=1.34.0
sympy~=1.12